+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ This document describes changes between each past release.
 1.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Remove template variable validation
 
 
 1.1.0 (2019-02-20)